Rollback renaming
