--- .editorconfig
@@ -1,30 +1,0 @@
-root = true
-
-[*]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-max_line_length = 100
-
-[{*.html,*.jinja}]
-indent_size = 4
-insert_final_newline = false
-max_line_length = off
-
-[*.json]
-insert_final_newline = false
-
-[*.md]
-trim_trailing_whitespace = false
-max_line_length = off
-
-[*.py]
-indent_size = 4
-
-[{*.xml,*.xsd}]
-indent_size = 4
-insert_final_newline = false
-max_line_length = off
