+++ mediux_posters/cli/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["settings_app"]
+
+from mediux_posters.cli.settings import app as settings_app
