+++ mediux_posters/mediux/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["CollectionSet", "File", "FileType", "Mediux", "MovieSet", "ShowSet"]
+
+from mediux_posters.mediux.schemas import CollectionSet, File, FileType, MovieSet, ShowSet
+from mediux_posters.mediux.service import Mediux
