+++ mediux_posters/services/_base/__init__.py
@@ -1,10 +1,10 @@
 __all__ = ["BaseCollection", "BaseEpisode", "BaseMovie", "BaseSeason", "BaseService", "BaseShow"]
 
-from visage.services._base.schemas import (
+from mediux_posters.services._base.schemas import (
     BaseCollection,
     BaseEpisode,
     BaseMovie,
     BaseSeason,
     BaseShow,
 )
-from visage.services._base.service import BaseService
+from mediux_posters.services._base.service import BaseService
