+++ mediux_posters/services/jellyfin/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Jellyfin"]
+
+from mediux_posters.services.jellyfin.service import Jellyfin
