+++ mediux_posters/services/jellyfin/schemas.py
@@ -6,8 +6,14 @@ from typing import Literal
 from pydantic import AliasPath, Field, field_validator
 from pydantic.alias_generators import to_pascal
 
-from visage.services._base import BaseCollection, BaseEpisode, BaseMovie, BaseSeason, BaseShow
-from visage.utils import BaseModel
+from mediux_posters.services._base import (
+    BaseCollection,
+    BaseEpisode,
+    BaseMovie,
+    BaseSeason,
+    BaseShow,
+)
+from mediux_posters.utils import BaseModel
 
 
 class JellyfinModel(BaseModel, alias_generator=to_pascal, extra="ignore"): ...
