+++ mediux_posters/services/jellyfin/service.py
@@ -11,12 +11,19 @@ from typing import Literal
 from httpx import Client, HTTPStatusError, RequestError, TimeoutException
 from pydantic import TypeAdapter, ValidationError
 
-from visage import __project__, __version__
-from visage.console import CONSOLE
-from visage.errors import AuthenticationError, ServiceError
-from visage.services._base import BaseService
-from visage.services.jellyfin.schemas import Collection, Episode, Library, Movie, Season, Show
-from visage.services.service_cache import ServiceCache
+from mediux_posters import __project__, __version__
+from mediux_posters.console import CONSOLE
+from mediux_posters.errors import AuthenticationError, ServiceError
+from mediux_posters.services._base import BaseService
+from mediux_posters.services.jellyfin.schemas import (
+    Collection,
+    Episode,
+    Library,
+    Movie,
+    Season,
+    Show,
+)
+from mediux_posters.services.service_cache import ServiceCache
 
 LOGGER = logging.getLogger(__name__)
