+++ mediux_posters/services/plex/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Plex"]
+
+from mediux_posters.services.plex.service import Plex
