+++ mediux_posters/services/plex/service.py
@@ -9,11 +9,11 @@ from platform import release, system
 from httpx import Client, HTTPStatusError, RequestError, TimeoutException
 from pydantic import TypeAdapter, ValidationError
 
-from visage import __project__, __version__
-from visage.console import CONSOLE
-from visage.errors import AuthenticationError, ServiceError
-from visage.services._base import BaseService
-from visage.services.plex.schemas import (
+from mediux_posters import __project__, __version__
+from mediux_posters.console import CONSOLE
+from mediux_posters.errors import AuthenticationError, ServiceError
+from mediux_posters.services._base import BaseService
+from mediux_posters.services.plex.schemas import (
     Collection,
     Episode,
     Library,
@@ -22,7 +22,7 @@ from visage.services.plex.schemas import (
     Season,
     Show,
 )
-from visage.services.service_cache import ServiceCache
+from mediux_posters.services.service_cache import ServiceCache
 
 LOGGER = logging.getLogger(__name__)
