+++ mediux_posters/settings.py
@@ -7,9 +7,9 @@ import tomli_w as tomlwriter
 from pydantic import BeforeValidator, Field
 from rich.panel import Panel
 
-from visage import get_config_root
-from visage.console import CONSOLE
-from visage.utils import BaseModel, blank_is_none, flatten_dict
+from mediux_posters import get_config_root
+from mediux_posters.console import CONSOLE
+from mediux_posters.utils import BaseModel, blank_is_none, flatten_dict
 
 try:
     from typing import Self  # Python >= 3.11
