+++ tests/jellyfin_test.py
@@ -5,7 +5,6 @@ from pathlib import Path
 
 import pytest
 from pytest_httpx import HTTPXMock
-
 from visage.services.jellyfin import Jellyfin
