--- visage/cli/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["settings_app"]
-
-from visage.cli.settings import app as settings_app
