--- visage/mediux/__init__.py
@@ -1,4 +1,0 @@
-__all__ = ["CollectionSet", "File", "FileType", "Mediux", "MovieSet", "ShowSet"]
-
-from visage.mediux.schemas import CollectionSet, File, FileType, MovieSet, ShowSet
-from visage.mediux.service import Mediux
