--- visage/services/jellyfin/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["Jellyfin"]
-
-from visage.services.jellyfin.service import Jellyfin
