--- visage/services/plex/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["Plex"]
-
-from visage.services.plex.service import Plex
